+++ tx-backend/src/main/resources/db/migration/R__create_contract_agreement_view.sql
@@ -1,4 +1,4 @@
-create or replace view contract_agreement_info_view as
+create or replace view contract_agreement_view as
 SELECT *
 FROM ((SELECT assets_as_built.id, contract_agreement_id, 'assets_as_built' as asset_type, created
        FROM assets_as_built