+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/contracts/ContractControllerIT.java
@@ -39,7 +39,7 @@ import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.eclipse.tractusx.traceability.common.security.JwtRole.ADMIN;
 
-class ContractsControllerIT extends IntegrationTestSpecification {
+class ContractControllerIT extends IntegrationTestSpecification {
 
     @Autowired
     AssetsSupport assetsSupport;